fix seed imports, move main styles to css
