+++ public/css/styles.css
@@ -1,0 +1,32 @@
+/* Add custom styles here */
+.custom-navbar {
+  margin: 20px;
+  border-radius: 10px;
+  height: 5rem;
+}
+.navbar-nav {
+  width: 100%;
+  display: flex;
+  justify-content: space-evenly;
+}
+.left-sidebar {
+  width: 20%;
+  background-color: #f5f5f5;
+  min-height: calc(100vh - 6rem);
+  position: fixed;
+  top: 9rem;
+  left: 0;
+}
+.grid-container {
+  margin-left: 20%;
+  padding: 6rem;
+  gap: 4rem;
+}
+.grid-item {
+  border: 1px solid #ccc;
+  padding: 10px;
+  min-height: calc(50vh - 12rem);
+}
+main {
+  margin-top: 9rem;
+}
