+++ seeds/index.js
@@ -1,5 +1,5 @@
 const sequelize = require("../config/connection");
-const seedFood = require("./paintingData");
+const seedFood = require("./foodData");
 
 const seedAll = async () => {
   await sequelize.sync({ force: true });
